Implement a keymap for PRK firware
[PRK firmware](https://github.com/picoruby/prk_firmware) runs on RP2040s
and supports duplexing out of the box, making it a perfect fit for the
Cheapino, especially if you write Ruby in your day job like I do.

`keymap.rb` implements [Peter Jang's beginner-friendly keyboard
layout](https://keymapdb.com/keymaps/peterxjang/) for the Cheapino on PRK.

The vital bit of this `keymap.rb` is the pin mapping for duplexing.